+++ pyoffers/models/affiliate_offer.py
@@ -1,0 +1,13 @@
+from .core import Model, ModelManager
+
+
+class AffiliateOffer(Model):
+    """"
+    Model that defines relation between Offer and Affiliate
+    """
+    pass
+
+
+class AffiliateOfferManager(ModelManager):
+    model = AffiliateOffer
+    name = 'affiliate_offers'